refactor: check user who can remove work

+++ server/controllers/work/deleteWork.js
@@ -4,13 +4,19 @@ const { paramsValidation } = require('../../utils/validation');
 
 const deleteWork = async (req, res, next) => {
   try {
-    const { id } = await paramsValidation.validateAsync(req.params);
-    const work = await Work.findByPk(id);
-    if (!work) {
-      throw customError('work does not exist', 400);
+    const { userId } = req.body;
+    const { providerID } = req;
+    if (userId !== providerID) {
+      throw customError('Unauthorized', 401);
+    } else {
+      const { id } = await paramsValidation.validateAsync(req.params);
+      const work = await Work.findByPk(id);
+      if (!work) {
+        throw customError('work does not exist', 400);
+      }
+      await work.destroy({ where: { id } });
+      res.json({ msg: 'work deleted successfully' });
     }
-    await work.destroy({ where: { id } });
-    res.json({ msg: 'work deleted successfully' });
   } catch (error) {
     if (error.name === 'ValidationError') {
       return next(customError(error.message, 400));
